feat(CoreUtils): move custom toggle to core (to allow reuse in other UI sections)
