+++ package.json
@@ -22,5 +22,5 @@
   },
   "unity": "2021.3",
   "unityRelease": "23f1",
-  "version": "0.218.2-dev"
+  "version": "0.219.0-dev"
 }
