+++ packagecore/CoreUtils.cs
@@ -480,5 +480,135 @@ namespace uk.novavoidhowl.dev.cvrfury.packagecore
       return true;
     }
   }
+
+  // Custom SwitchToggle class that creates a mobile-style toggle switch
+  public class SwitchToggle : BindableElement, INotifyValueChanged<bool>
+  {
+    private VisualElement track;
+    private VisualElement knob;
+    private Label label;
+    private bool _value;
+    private SerializedProperty _boundProperty;
+
+    public bool value
+    {
+      get { return _value; }
+      set
+      {
+        if (_value != value)
+        {
+          bool previousValue = _value;
+          _value = value;
+          UpdateVisualState();
+          using (ChangeEvent<bool> evt = ChangeEvent<bool>.GetPooled(previousValue, _value))
+          {
+            evt.target = this;
+            SendEvent(evt);
+          }
+        }
+      }
+    }
+
+    // Event that clients can subscribe to
+    public event EventCallback<ChangeEvent<bool>> onValueChanged;
+
+    public SwitchToggle(string labelText = null)
+    {
+      AddToClassList("switch-toggle");
+
+      // Create container for the label and toggle
+      var container = new VisualElement();
+      container.AddToClassList("switch-toggle-container");
+      hierarchy.Add(container);
+
+      // Add label if provided
+      if (!string.IsNullOrEmpty(labelText))
+      {
+        label = new Label(labelText);
+        label.AddToClassList("switch-toggle-label");
+        container.Add(label);
+      }
+
+      // Create toggle track
+      track = new VisualElement();
+      track.AddToClassList("switch-toggle-track");
+      container.Add(track);
+
+      // Create toggle knob
+      knob = new VisualElement();
+      knob.AddToClassList("switch-toggle-knob");
+      track.Add(knob);
+
+      // Register callbacks
+      RegisterCallback<ClickEvent>(OnClick);
+      RegisterCallback<AttachToPanelEvent>(OnAttachToPanel);
+    }
+
+    private void OnAttachToPanel(AttachToPanelEvent evt)
+    {
+      UpdateVisualState();
+    }
+
+    private void OnClick(ClickEvent evt)
+    {
+      value = !_value;
+      evt.StopPropagation();
+    }
+
+    private void UpdateVisualState()
+    {
+      if (_value)
+      {
+        track.AddToClassList("switch-toggle-track-active");
+        knob.AddToClassList("switch-toggle-knob-active");
+      }
+      else
+      {
+        track.RemoveFromClassList("switch-toggle-track-active");
+        knob.RemoveFromClassList("switch-toggle-knob-active");
+      }
+    }
+
+    public void RegisterValueChangedCallback(EventCallback<ChangeEvent<bool>> callback)
+    {
+      onValueChanged += callback;
+      RegisterCallback<ChangeEvent<bool>>(callback);
+    }
+
+    public void UnregisterValueChangedCallback(EventCallback<ChangeEvent<bool>> callback)
+    {
+      onValueChanged -= callback;
+      UnregisterCallback<ChangeEvent<bool>>(callback);
+    }
+
+    // Implementation of INotifyValueChanged<bool>
+    public void SetValueWithoutNotify(bool newValue)
+    {
+      if (_value != newValue)
+      {
+        _value = newValue;
+        UpdateVisualState();
+      }
+    }
+
+    // Add method to bind to a SerializedProperty
+    public void BindProperty(SerializedProperty property)
+    {
+      _boundProperty = property;
+      if (_boundProperty != null)
+      {
+        SetValueWithoutNotify(_boundProperty.boolValue);
+      }
+    }
+
+    // Method to update value from bound property
+    public void UpdateFromProperty()
+    {
+      if (_boundProperty != null && _boundProperty.serializedObject != null)
+      {
+        SetValueWithoutNotify(_boundProperty.boolValue);
+      }
+    }
+  }
 #endif
 }
